fix clang-tidy error

+++ tt_metal/fabric/compressed_routing_path.cpp
@@ -17,7 +17,7 @@ void intra_mesh_routing_path_t<1, false>::calculate_chip_to_all_routing_fields(
     uint64_t* route_ptr = reinterpret_cast<uint64_t*>(&paths);
     route_ptr[0] = 0;
     for (uint16_t hops = 1; hops < num_chips; ++hops) {
-        const uint64_t shift_amount = static_cast<uint64_t>((hops - 1) * FIELD_WIDTH);
+        const uint64_t shift_amount = static_cast<uint64_t>(hops - 1) * static_cast<uint64_t>(FIELD_WIDTH);
         route_ptr[hops] =
             (FWD_ONLY_FIELD & ((1ULL << shift_amount) - 1ULL)) | (static_cast<uint64_t>(WRITE_ONLY) << shift_amount);
     }
